# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"URL": {
				"type": "string",
				"example": "https://example.com/example",
				"format": "uri",
				"description": "URL format.",
				"pattern": "^(https?:\\/\\/)?(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)$"
			},
			"LongString": {
				"type": "string",
				"minLength": 2,
				"maxLength": 300
			},
			"GetItemPicUrlOut": {
				"properties": {
					"itemPicUrl": {
						"$ref": "#/components/schemas/URL"
					},
					"itemPicKey": {
						"$ref": "#/components/schemas/LongString"
					}
				},
				"required": [
					"itemPicUrl",
					"itemPicKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorDetail": {
				"properties": {
					"field": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForbiddenError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"S3ValidationError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UUID": {
				"type": "string",
				"example": "52907745-7672-470e-a803-a2f8feb52944",
				"format": "uuid",
				"description": "Stringified UUIDv4.\r\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
			"ImageFilename": {
				"type": "string",
				"example": "my-image.jpg",
				"format": "image-filename",
				"description": "Valid filename for images with allowed extensions.\r\nContains only English letters, numbers, hyphens, and underscores.\r\nMust end with a valid image extension (.jpg, .jpeg, .png, .webp, .svg, .heic).",
				"pattern": "^[a-zA-Z0-9-_]+\\.(jpg|jpeg|png|webp|svg|heic)$",
				"minLength": 5,
				"maxLength": 255
			},
			"GetItemPicUrlIn": {
				"properties": {
					"restaurantId": {
						"$ref": "#/components/schemas/UUID"
					},
					"branchId": {
						"$ref": "#/components/schemas/UUID"
					},
					"fileName": {
						"$ref": "#/components/schemas/ImageFilename"
					}
				},
				"required": [
					"restaurantId",
					"branchId",
					"fileName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetRestaurantCoverPicUrlOut": {
				"properties": {
					"restaurantCoverPicUrl": {
						"$ref": "#/components/schemas/URL"
					},
					"restaurantCoverPicKey": {
						"$ref": "#/components/schemas/LongString"
					}
				},
				"required": [
					"restaurantCoverPicUrl",
					"restaurantCoverPicKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetRestaurantPicUrlIn": {
				"properties": {
					"restaurantId": {
						"$ref": "#/components/schemas/UUID"
					},
					"fileName": {
						"$ref": "#/components/schemas/ImageFilename"
					}
				},
				"required": [
					"restaurantId",
					"fileName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetRestaurantAvatarPicUrlOut": {
				"properties": {
					"restaurantAvatarPicUrl": {
						"$ref": "#/components/schemas/URL"
					},
					"restaurantAvatarPicKey": {
						"$ref": "#/components/schemas/LongString"
					}
				},
				"required": [
					"restaurantAvatarPicUrl",
					"restaurantAvatarPicKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetRestaurantLogoPicUrlOut": {
				"properties": {
					"restaurantLogoPicUrl": {
						"$ref": "#/components/schemas/URL"
					},
					"restaurantLogoPicKey": {
						"$ref": "#/components/schemas/LongString"
					}
				},
				"required": [
					"restaurantLogoPicUrl",
					"restaurantLogoPicKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetMenuFaviconUrlOut": {
				"properties": {
					"menuFaviconUrl": {
						"$ref": "#/components/schemas/URL"
					},
					"menuFaviconKey": {
						"$ref": "#/components/schemas/LongString"
					}
				},
				"required": [
					"menuFaviconUrl",
					"menuFaviconKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetMenuFaviconUrlIn": {
				"properties": {
					"restaurantId": {
						"$ref": "#/components/schemas/UUID"
					},
					"branchId": {
						"$ref": "#/components/schemas/UUID"
					},
					"menuId": {
						"$ref": "#/components/schemas/UUID"
					},
					"fileName": {
						"$ref": "#/components/schemas/ImageFilename"
					}
				},
				"required": [
					"restaurantId",
					"branchId",
					"menuId",
					"fileName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"SessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session-id"
			},
			"JWTAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access-token"
			}
		}
	},
	"info": {
		"title": "Menuchi-API",
		"version": "1.0.0",
		"description": "Menuchi API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Mohammad Hosseini"
		}
	},
	"paths": {
		"/s3/get-item-pic-url": {
			"post": {
				"operationId": "GeneratePutItemPicPresignedUrl",
				"responses": {
					"201": {
						"description": "Item pic presigned url generated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetItemPicUrlOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4224 S3ValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/S3ValidationError"
								}
							}
						}
					}
				},
				"description": "Generates a pre-signed URL for uploading an item picture to S3.",
				"tags": [
					"S3"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetItemPicUrlIn"
							}
						}
					}
				}
			}
		},
		"/s3/get-res-cover-pic-url": {
			"post": {
				"operationId": "GeneratePutRestaurantCoverPicPresignedUrl",
				"responses": {
					"201": {
						"description": "Restaurant cover pic presigned url generated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRestaurantCoverPicUrlOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4224 S3ValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/S3ValidationError"
								}
							}
						}
					}
				},
				"description": "Generates a pre-signed URL for uploading an restaurant cover picture to S3.",
				"tags": [
					"S3"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetRestaurantPicUrlIn"
							}
						}
					}
				}
			}
		},
		"/s3/get-res-avatar-pic-url": {
			"post": {
				"operationId": "GeneratePutRestaurantAvatarPicPresignedUrl",
				"responses": {
					"201": {
						"description": "Restaurant avatar pic presigned url generated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRestaurantAvatarPicUrlOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4224 S3ValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/S3ValidationError"
								}
							}
						}
					}
				},
				"description": "Generates a pre-signed URL for uploading an restaurant avatar picture to S3.",
				"tags": [
					"S3"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetRestaurantPicUrlIn"
							}
						}
					}
				}
			}
		},
		"/s3/get-res-logo-pic-url": {
			"post": {
				"operationId": "GeneratePutRestaurantLogoPicPresignedUrl",
				"responses": {
					"201": {
						"description": "Restaurant logo pic presigned url generated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetRestaurantLogoPicUrlOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4224 S3ValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/S3ValidationError"
								}
							}
						}
					}
				},
				"description": "Generates a pre-signed URL for uploading an restaurant logo picture to S3.",
				"tags": [
					"S3"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetRestaurantPicUrlIn"
							}
						}
					}
				}
			}
		},
		"/s3/get-menu-favicon-url": {
			"post": {
				"operationId": "GeneratePutMenuFaviconPresignedUrl",
				"responses": {
					"201": {
						"description": "Menu favicon presigned url generated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetMenuFaviconUrlOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"422": {
						"description": "4224 S3ValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/S3ValidationError"
								}
							}
						}
					}
				},
				"description": "Generates a pre-signed URL for uploading an menu favicon to S3.",
				"tags": [
					"S3"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetMenuFaviconUrlIn"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}